Fixes for game updates ^1.93.1; Update v0.6.3
Improvements to the build script:
Regular expressions can now be generated and used automatically just by providing raw code segments of the original and the modified code;
Dictionary expressions can be generated with a similar mechanism.

Other changes:
The settings controls are now programmatically generated;
Removed the custom font setting because this was added to the base game; "Trebuchet MS" is now the default font for that setting instead of "sans-serif";
Removed the "hide all links" setting (the "Hide Links" option was removed from the base game), this could be replaced with a new "hide links" setting.